+++ aksel.nav.no/website/components/website-modules/search/parts/HitCollection.tsx
@@ -74,8 +74,8 @@ export function Collection({
   return (
     <div>
       {heading && tag && (
-        <h3
-          className={cl("sticky top-0 z-10 p-2 px-4 md:px-10", {
+        <h2
+          className={cl("sticky top-0 z-10 p-2 px-6 md:px-10", {
             "bg-deepblue-50": ["komponent_artikkel", "ds_artikkel"].includes(
               tag,
             ),
@@ -87,18 +87,18 @@ export function Collection({
           })}
         >
           <Tag hTag="span" type={tag} size="small" inline count={hits.length} />
-        </h3>
+        </h2>
       )}
       {heading && !tag && (
         <Heading
-          className="sticky top-0 z-10 bg-surface-subtle p-2 px-4  md:px-10"
+          className="sticky top-0 z-10 bg-surface-subtle p-2 px-6  md:px-10"
           size="small"
-          level="3"
+          level="2"
         >
           {heading}
         </Heading>
       )}
-      <ul className="mt-2 px-0 md:px-6">
+      <ul className="px-2 md:px-6">
         {split.initial.map((x, xi) => (
           <Hit
             key={xi}