+++ aksel.nav.no/website/pages/eksempler/bodylong/override-tag.tsx
@@ -14,5 +14,5 @@ export const Demo = {
 };
 
 export const args = {
-  index: 8,
+  index: 6,
 };