+++ aksel.nav.no/website/pages/eksempler/detail/uppercase.tsx
@@ -1,0 +1,22 @@
+import { Detail } from "@navikt/ds-react";
+import { withDsExample } from "@/web/examples/withDsExample";
+
+const Example = () => {
+  return (
+    <Detail uppercase>
+      Du må gjøre en filtrering for å se brukere i listen.
+    </Detail>
+  );
+};
+
+// EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
+export default withDsExample(Example, { variant: "full" });
+
+/* Storybook story */
+export const Demo = {
+  render: Example,
+};
+
+export const args = {
+  index: 1,
+};